requirements.txt
flask
flask-cors
requests
python-dotenv
google-genai
psycopg2-binary
sqlalchemy
pubchempy
